+++ toxsign/templates/assays/partial_subfactor_show.html
@@ -31,7 +31,9 @@
                         <th scope="col">Chemical</th>
                         <th scope="col">Dose</th>
                         <th scope="col">Exposure</th>
+                        {% if can_edit %}
                         <th scope="col">Action</th>
+                        {% endif %}
                     </tr>
                 </thead>
                 <tbody>
@@ -41,17 +43,19 @@
                             <td>{% if chemical.chemical %}{{chemical.chemical.name}}{% else %}{{chemical.chemical_slug}}{% endif %}</td>
                             <td>{% if chemical.dose_value %} {{chemical.dose_value}} {{chemical.dose_unit}}  {% else %} No dose provided {% endif %}</td>
                             <td>{% if chemical.exposure_time %} {{chemical.exposure_time}} hours {% else %} No exposure time provided {% endif %}</td>
+                            {% if can_edit %}
                             <td>
                                  <a role="button" href="{% url 'assays:chemical_subfactor_edit' chemical.id %}" class="btn btn-info"><i class="fas fa-edit"></i></a>
                                 <a role="button" href="{% url 'assays:subfactor_delete' 'chemical' chemical.id %}" onclick="return confirm('Are you sure you want to delete this subfactor?');" class="btn btn-danger"><i class="fas fa-trash"></i></a>
                             </td>
+                            {% endif %}
                         </tr>
                     {% endfor %}
                 </tbody>
             </table>
-            
+            {% if can_edit %}
             <a role="button" href="{% url 'assays:chemical_subfactor_create' factor.tsx_id %}"class="btn btn-info float-right"><i class="fas fa-external-link-alt"></i>New</a></p>
-            
+            {% endif %}
         </div>
         <div class="tab-pane" id="biologicalmodal">
             <p>No biological subfactors linked to this factor</p>
